refactor: 패키지 리팩터링
사용하지않은 객체 삭제 및 객체 패키지 리팩터링
