+++ src/test/java/com/poortorich/user/fixture/enums/NameValidationCase.java
@@ -1,7 +1,7 @@
 package com.poortorich.user.fixture.enums;
 
 
-import com.poortorich.global.TestCase;
+import com.poortorich.global.testcases.TestCase;
 import com.poortorich.user.constants.UserResponseMessages;
 
 public enum NameValidationCase implements TestCase<String, String> {
